Add javadoc to the BitUtils

+++ src/main/java/core/bit/BitUtils.java
@@ -1,5 +1,14 @@
 package core.bit;
 
+/**
+ * Class that offers utils methods for
+ * boolean type variables. It can be used
+ * to perform a standard logical operations
+ * in functional style.
+ *
+ * @see    Bit
+ * @author Iliya Zakharchenia
+ */
 public abstract class BitUtils {
     public final static String DEFAULT_NAME = "UNNAMED";
     public final static String FALSE_STRING = "0";
@@ -69,17 +78,35 @@ public abstract class BitUtils {
         return false;
     }
 
+    /**
+     * This method is building an integer number
+     * from bit varargs.
+     * <br>
+     * Example: int x = BitUtils.numberFromBitSequence(1, 0, 1);
+     * <br>
+     * // x == 5 ('101' in 2ns)
+     *
+     */
     public static int numberFromBitSequence(boolean... bits) {
         int sum = 0;
         int len = bits.length;
 
         for (int i = len-1; i > -1; i = i-1) {
-            if (bits[i]) sum += Math.pow(2, len-i-1);
+            if (bits[i]) sum += (int) Math.pow(2, len-i-1);
         }
 
         return sum;
     }
 
+    /**
+     * This method is building a 2ns-number
+     * from int value.
+     * <br>
+     * Example: boolean[] x = BitUtils.bitSequenceFromNumber(5);
+     * <br>
+     * // x == [true, false, true] ('101' in 2ns)
+     *
+     */
     public static boolean[] bitSequenceFromNumber(int number) {
         if (number < 0)
             throw new IllegalArgumentException("Number shouldn't been less then 0!");
@@ -96,13 +123,18 @@ public abstract class BitUtils {
             if (value % 2 == 0) sequence[i] = false;
             else {
                 sequence[i] = true;
-                different -= Math.pow(2, length-i-1);
+                different -= (int) Math.pow(2, length-i-1);
             }
         }
 
         return sequence;
     }
 
+    /**
+     * This method is adding leading zeroes to
+     * the actual sequence.
+     *
+     */
     public static boolean[] withLeadingZeroes(int quantityOfZeroes, boolean[] sequence) {
         if (quantityOfZeroes < 0)
             throw new IllegalArgumentException("Quantity of leading zeroes shouldn't be less then 0!");
